+++ frontend/mock/azureMsalBrowser.js
