+++ frontend/src/pages/Admin.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useRef } from 'react';
 import './Home.css';
-import { getAdminData } from '../components/api';
+import { getAdminData } from '../api/api';
 import { useMsal } from '@azure/msal-react';
 import appInsights from '../components/appInsights';
 import Loading, { sleep } from '../components/Loading'; // Import the Loading component
