+++ frontend/vite.config.js
@@ -47,11 +47,16 @@ console.log(`Mocking ${isMockEnabled ? 'enabled' : 'disabled'}`);
 const getAliases = () => {
   if (isMockEnabled) {
     console.log('Redirecting MSAL imports to mock implementation');
+    // !! Important : This does the actual mocking in a transparent way
     return {
-      '@azure/msal-browser': resolve(__dirname, 'src/mock/mockAzureMsalBrowser.js'),
+      '@azure/msal-browser': resolve(__dirname, 'mock/azureMsalBrowser.js'),
+      '@/api/graphApi': resolve(__dirname, 'mock/graphApi.js'),
+      '@': resolve(__dirname, 'src')
     };
   }
-  return {};
+  return {
+    '@': resolve(__dirname, 'src')
+  };
 };
 
 export default defineConfig({
